Remove the .eps extension that is now added to jpg files.
The dvips l3backend is used by default and the only image extensions
that are allowed are ps and eps. So if any other file type is provided,
then the eps extension is tacked onto the end.  So this just removes the
extension.  Perhaps a custom backend should be used instead?