Fix cmex fonts.
The tex256.enc glyphs are completely wrong for them. So a direct mapping
is used instead of trying to get the encoding from a file somewhere.

Also, the glyph for the angbracketleftbig character in the truetype cmex
fonts is a soft hyphen which is a character that browsers will not show.
As such the character will not be seen if used. For example, if
`\big\langle` is used in TeX. Note that `\left\langle` could result in
this character as well. So the mapping is for the glyphs in the opentype
font instead of in the truetype font.

Note that the bakoma cmex otf fonts have invalid out of range glyphs in
them. So those glyphs need to be removed.